normalized simulation
